# Conflicts:
#	frontend/src/app/routing/routing.module.ts

+++ backend/routes/routes/institutionsRoutes.js
@@ -8,59 +8,94 @@ var router = express.Router({mergeParams: true});
 var defineRoutes = router =>{
 
     router.get('/:id/cms', async (req,res)=> {
-        let result = await InstitutionController.cms(req.params.id);
-        res.status(result?200:400).send(result);
+        let result = {};
+        if(!req.params.id)
+            result= {error:true,description:'you do\'nt have validation'};
+        else
+            result = await InstitutionController.cms(req.params.id);
+        res.status(result.error?400:200).send(result);
     });
 
     router.get('/getschoolsgs', async function(req,res){
         let result = await InstitutionController.getSchoolsGB();
-        res.status(result?200:400).send(result);
+        res.status(result.error?400:200).send(result);
     });
-    // TODO: need to check all the parameters for all!!!!!
     router.get('/:id', async function(req,res){
-        let result = await InstitutionController.getInstitution(req.params.id);
-        res.status(result?200:400).send(result);
+        let result = {};
+        if(!req.params.id)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.getInstitution(req.params.id);
+        res.status(result.error?400:200).send(result);
     });
 
     router.post('', checkAuth, async function(req,res){
-        let result =  await InstitutionController.createInstitution(req.body);
-        res.status(result?201:400).send(result);
+        let result = {};
+        if(!req.body.name || !req.body.suffix || !req.body.address || !req.body.geolocation || !req.body.image)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result =  await InstitutionController.createInstitution(req.body);
+        res.status(result.error?400:201).send(result);
     });
 
     router.get('', async function(req,res){
-        let result =  await InstitutionController.getInstitutionCollection(req.body);
-        res.status(result?200:400).send(result);
+        let result =  await InstitutionController.getInstitutionCollection();
+        res.status(result.error?400:200).send(result);
     });
 
     router.delete('', checkAuth, async function(req,res){
-        let result =  await InstitutionController.deleteInstitution(req.body._id);
-        res.status(200).send(result);
+        let result = {};
+        if(!req.body._id)
+            result =  {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.deleteInstitution(req.body._id);
+        res.status(result.error?400:200).send(result);
     });
 
     router.put('', checkAuth, async function(req,res){
-        let result =  await InstitutionController.updateInstitution(req.body);
-        res.status(200).send(result);
+        let result = {};
+        if(!req.body._id)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.updateInstitution(req.body);
+        res.status(result.error?400:200).send(result);
     });
 
     router.post('/addschool', checkAuth, async function(req,res){
-        let result =  await InstitutionController.addSchool(req.body);
-        res.status(result?201:400).send(result);
+        let result = {};
+        if(!req.body.institutionid || !req.body.schoolid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result =  await InstitutionController.addSchool(req.body);
+        res.status(result.error?400:201).send(result);
     });
 
 
     router.get('/getschools/:id', async function(req,res){
-        let result = await InstitutionController.getSchools(req.params.id);
-        res.status(result?200:400).send(result);
+        let result = {};
+        if(!req.params.id)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.getSchools(req.params.id);
+        res.status(result.error?400:200).send(result);
     });
 
     router.post('/permission', checkAuth, async function(req,res){
-        let result = await InstitutionController.addpermission(req.body);
-        res.status(result?200:400).send(result);
+        let result = {};
+        if(!req.body.institutionid || !req.body.userid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.addpermission(req.body);
+        res.status(result.error?400:200).send(result);
     });
 
     router.delete('/permission', checkAuth, async function(req,res){
-        let result = await InstitutionController.deletepermission(req.body);
-        res.status(result?200:400).send(result);
+        let result = {};
+        if(!req.body.institutionid || !req.body.userid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await InstitutionController.deletepermission(req.body);
+        res.status(result.error?400:200).send(result);
     });
 
     return router;
